[ast] Merge ExprKind priorities with AnyExpr definitions
