--- stevia_ast/src/expr_kind.rs
@@ -1,240 +1,0 @@
-/// Represents kind of an expression.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum ExprKind {
-    /// The if-then-else expression kind.
-    IfThenElse,
-    /// The symbol expression kind
-    Symbol,
-
-    /// The constant boolean expression kind
-    BoolConst,
-    /// The boolean equality (also known as if-and-only-if) expression kind
-    BoolEquals,
-    /// The not boolean expression kind
-    Not,
-    /// The and boolean expression kind
-    And,
-    /// The or boolean expression kind
-    Or,
-    /// The implies boolean expression kind
-    Implies,
-    /// The xor (either-or) expression kind
-    Xor,
-
-    /// The constant bitvec term expression kind
-    BitvecConst,
-    // /// The bitvec equality term expression kind
-    BitvecEquals,
-    /// The bitvec negation term expression kind
-    Neg,
-    /// The bitvec add term expression kind
-    Add,
-    /// The bitvec mul term expression kind
-    Mul,
-    /// The bitvec sub term expression kind
-    Sub,
-    /// The bitvec udiv (unsigned division) term expression kind
-    UnsignedDiv,
-    /// The bitvec sdiv (signed division) term expression kind
-    SignedDiv,
-    /// The bitvec smod (signed remainder + sign match) term expression kind
-    SignedModulo,
-    /// The bitvec urem (unsigned remainder) term expression kind
-    UnsignedRemainder,
-    /// The bitvec srem (signed remainder) term expression kind
-    SignedRemainder,
-
-    /// The bitwise-not term expression kind
-    BitNot,
-    /// The bitwise-and term expression kind
-    BitAnd,
-    /// The bitwise-or term expression kind
-    BitOr,
-    /// The bitwise-xor term expression kind
-    BitXor,
-
-    /// The signed greater-than-or-equals term expression kind
-    SignedGreaterEquals,
-    /// The signed greater-than term expression kind
-    SignedGreaterThan,
-    /// The signed less-than-or-equals term expression kind
-    SignedLessEquals,
-    /// The signed less-than term expression kind
-    SignedLessThan,
-    /// The unsigned greater-than-or-equals term expression kind
-    UnsignedGreaterEquals,
-    /// The unsigned greater-than term expression kind
-    UnsignedGreaterThan,
-    /// The unsigned less-than-or-equals term expression kind
-    UnsignedLessEquals,
-    /// The unsigned less-than term expression kind
-    UnsignedLessThan,
-
-    /// The shift-left term expression kind
-    ShiftLeft,
-    /// The logical shift-right term expression kind
-    LogicalShiftRight,
-    /// The arithmetic shift-right term expression kind
-    ArithmeticShiftRight,
-
-    /// The bitvec concatenate term expression kind
-    Concat,
-    /// The bitvec extraction term expression kind
-    Extract,
-    /// The bitvec sign-extension term expression kind
-    SignExtend,
-    /// The bitvec zero-extension term expression kind
-    ZeroExtend,
-
-    /// The array-read expression kind
-    ArrayRead,
-    /// The array-write expression kind
-    ArrayWrite
-}
-
-impl ExprKind {
-    /// Returns the camel-case name of the expression kind.
-    pub fn camel_name(self) -> &'static str {
-        use self::ExprKind::*;
-        match self {
-            IfThenElse => "IfThenElse",
-            Symbol => "Symbol",
-            BoolConst => "BoolConst",
-            BoolEquals => "BoolEquals",
-            Not => "Not",
-            And => "And",
-            Or => "Or",
-            Implies => "Implies",
-            Xor => "Xor",
-            BitvecConst => "BitvecConst",
-            BitvecEquals => "BitvecEquals",
-            Neg => "Neg",
-            Add => "Add",
-            Mul => "Mul",
-            Sub => "Sub",
-            UnsignedDiv => "UnsignedDiv",
-            SignedDiv => "SignedDiv",
-            SignedModulo => "SignedModulo",
-            UnsignedRemainder => "UnsignedRemainder",
-            SignedRemainder => "SignedRemainder",
-            BitNot => "BitNot",
-            BitAnd => "BitAnd",
-            BitOr => "BitOr",
-            BitXor => "BitXor",
-            SignedGreaterEquals => "SignedGreaterEquals",
-            SignedGreaterThan => "SignedGreaterThan",
-            SignedLessEquals => "SignedLessEquals",
-            SignedLessThan => "SignedLessThan",
-            UnsignedGreaterEquals => "UnsignedGreaterEquals",
-            UnsignedGreaterThan => "UnsignedGreaterThan",
-            UnsignedLessEquals => "UnsignedLessEquals",
-            UnsignedLessThan => "UnsignedLessThan",
-            ShiftLeft => "ShiftLeft",
-            LogicalShiftRight => "LogicalShiftRight",
-            ArithmeticShiftRight => "ArithmeticShiftRight",
-            Concat => "Concat",
-            Extract => "Extract",
-            SignExtend => "SignExtend",
-            ZeroExtend => "ZeroExtend",
-            ArrayRead => "ArrayRead",
-            ArrayWrite => "ArrayWrite"
-        }
-    }
-}
-
-/// This trait should be implemented by all expressions and structures that
-/// represent an expression kind.
-/// 
-/// This is obviously true for `ExprKind` itself but also for all concrete expression types.
-pub trait HasKind {
-    /// Returns the kind of `self`.
-    fn kind(&self) -> ExprKind;
-}
-
-impl HasKind for ExprKind {
-    fn kind(&self) -> ExprKind {
-        *self
-    }
-}
-
-/// The priority of an expression kind.
-/// 
-/// This is used in order to improve normalization of symmetric
-/// expression by sorting their child expressions in a static ordering.
-/// 
-/// The priority of an expression kind decides this ordering.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct Priority(u32);
-
-/// Expression kinds do need to implement this trait in order
-/// to improve expression normalization.
-pub trait HasPriority {
-    /// Returns the priority of `self`.
-    fn priority(&self) -> Priority;
-}
-
-const BASE_PRIORITY_FORMULA: u32    = 100;
-const BASE_PRIORITY_ARITHMETIC: u32 = 200;
-const BASE_PRIORITY_BITWISE: u32    = 300;
-const BASE_PRIORITY_COMPARISON: u32 = 400;
-const BASE_PRIORITY_SHIFT: u32      = 500;
-const BASE_PRIORITY_CAST: u32       = 600;
-const BASE_PRIORITY_ARRAY: u32      = 700;
-const BASE_PRIORITY_GENERIC: u32    = 800;
-
-impl HasPriority for ExprKind {
-    fn priority(&self) -> Priority {
-        use self::ExprKind::*;
-        let prio_val = match *self {
-            IfThenElse => BASE_PRIORITY_GENERIC,
-            Symbol     => BASE_PRIORITY_GENERIC + 1,
-
-            BoolConst  => BASE_PRIORITY_FORMULA,
-            BoolEquals => BASE_PRIORITY_FORMULA + 1,
-            Not        => BASE_PRIORITY_FORMULA + 2,
-            And        => BASE_PRIORITY_FORMULA + 3,
-            Or         => BASE_PRIORITY_FORMULA + 4,
-            Implies    => BASE_PRIORITY_FORMULA + 5,
-            Xor        => BASE_PRIORITY_FORMULA + 6,
-
-            BitvecConst       => BASE_PRIORITY_ARITHMETIC,
-            BitvecEquals      => BASE_PRIORITY_ARITHMETIC +  1,
-            Neg               => BASE_PRIORITY_ARITHMETIC +  2,
-            Add               => BASE_PRIORITY_ARITHMETIC +  3,
-            Mul               => BASE_PRIORITY_ARITHMETIC +  4,
-            Sub               => BASE_PRIORITY_ARITHMETIC +  5,
-            UnsignedDiv       => BASE_PRIORITY_ARITHMETIC +  6,
-            SignedDiv         => BASE_PRIORITY_ARITHMETIC +  7,
-            SignedModulo      => BASE_PRIORITY_ARITHMETIC +  8,
-            UnsignedRemainder => BASE_PRIORITY_ARITHMETIC +  9,
-            SignedRemainder   => BASE_PRIORITY_ARITHMETIC + 10,
-
-            BitNot => BASE_PRIORITY_BITWISE,
-            BitAnd => BASE_PRIORITY_BITWISE + 1,
-            BitOr  => BASE_PRIORITY_BITWISE + 2,
-            BitXor => BASE_PRIORITY_BITWISE + 3,
-
-            SignedGreaterEquals   => BASE_PRIORITY_COMPARISON,
-            SignedGreaterThan     => BASE_PRIORITY_COMPARISON + 1,
-            SignedLessEquals      => BASE_PRIORITY_COMPARISON + 2,
-            SignedLessThan        => BASE_PRIORITY_COMPARISON + 3,
-            UnsignedGreaterEquals => BASE_PRIORITY_COMPARISON + 4,
-            UnsignedGreaterThan   => BASE_PRIORITY_COMPARISON + 5,
-            UnsignedLessEquals    => BASE_PRIORITY_COMPARISON + 6,
-            UnsignedLessThan      => BASE_PRIORITY_COMPARISON + 7,
-
-            ShiftLeft            => BASE_PRIORITY_SHIFT,
-            LogicalShiftRight    => BASE_PRIORITY_SHIFT + 1,
-            ArithmeticShiftRight => BASE_PRIORITY_SHIFT + 2,
-
-            Concat     => BASE_PRIORITY_CAST,
-            Extract    => BASE_PRIORITY_CAST + 1,
-            SignExtend => BASE_PRIORITY_CAST + 2,
-            ZeroExtend => BASE_PRIORITY_CAST + 3,
-
-            ArrayRead  => BASE_PRIORITY_ARRAY,
-            ArrayWrite => BASE_PRIORITY_ARRAY + 1
-        };
-        Priority(prio_val)
-    }
-}
